[IMP] purchase{_stock}: common tour helpers
Helpers for tours were written to ease tours regarding PO and product
catalog in two different branches targeting different versions.
This commit moves helpers into one file and organize/rename them.
